@@ -5,9 +5,14 @@ class LastThreeCommentsDecorator
     @presenter = presenter
   end
 
-  def as_json(options={})
+  def as_json(_options={})
+    current_user = @presenter.current_user
     @presenter.as_json.tap do |post|
-      post[:interactions].merge!(:comments => CommentPresenter.as_collection(@presenter.post.last_three_comments))
+      post[:interactions].merge!(comments: CommentPresenter.as_collection(
+        @presenter.post.last_three_comments,
+        :as_json,
+        current_user
+      ))
     end
   end
 end
