@@ -16,8 +16,23 @@ class LikesController < ApplicationController
     authenticate_user!
   end
 
+  def index
+    like = if params[:post_id]
+             like_service.find_for_post(params[:post_id])
+           else
+             like_service.find_for_comment(params[:comment_id])
+           end
+    render json: like
+      .includes(author: :profile)
+      .as_api_response(:backbone)
+  end
+
   def create
-    like = like_service.create(params[:post_id])
+    like = if params[:post_id]
+             like_service.create_for_post(params[:post_id])
+           else
+             like_service.create_for_comment(params[:comment_id])
+           end
   rescue ActiveRecord::RecordNotFound, ActiveRecord::RecordInvalid
     render plain: I18n.t("likes.create.error"), status: 422
   else
@@ -34,12 +49,6 @@ class LikesController < ApplicationController
     else
       render plain: I18n.t("likes.destroy.error"), status: 404
     end
-  end
-
-  def index
-    render json: like_service.find_for_post(params[:post_id])
-      .includes(author: :profile)
-      .as_api_response(:backbone)
   end
 
   private
