@@ -26,7 +26,7 @@ module MobileHelper
   end
 
   def mobile_like_icon(post)
-    if current_user && current_user.liked?(post)
+    if current_user&.liked?(post)
       link_to content_tag(:span, post.likes.size, class: "count like-count"),
               "#",
               data:  {url: post_like_path(post.id, current_user.like_for(post).id)},
@@ -35,6 +35,20 @@ module MobileHelper
       link_to content_tag(:span, post.likes.size, class: "count like-count"),
               "#",
               data:  {url: post_likes_path(post.id)},
+              class: "entypo-heart like-action inactive"
+    end
+  end
+
+  def mobile_like_comment_icon(comment)
+    if current_user&.liked?(comment)
+      link_to content_tag(:span, comment.likes.size, class: "count like-count"),
+              "#",
+              data:  {url: comment_like_path(comment.id, current_user.like_for(comment).id)},
+              class: "entypo-heart like-action active"
+    else
+      link_to content_tag(:span, comment.likes.size, class: "count like-count"),
+              "#",
+              data:  {url: comment_likes_path(comment.id)},
               class: "entypo-heart like-action inactive"
     end
   end
