@@ -16,6 +16,8 @@ module NotificationsHelper
       elsif %w(Notifications::CommentOnPost Notifications::AlsoCommented Notifications::Reshared Notifications::Liked)
             .include?(note.type)
         opts.merge!(opts_for_post(note.linked_object))
+      elsif note.is_a?(Notifications::LikedComment)
+        opts.merge!(opts_for_comment(note.linked_object))
       elsif note.is_a?(Notifications::ContactsBirthday)
         opts.merge!(opts_for_birthday(note))
       end
@@ -33,7 +35,16 @@ module NotificationsHelper
       post_link:   link_to(post_page_title(post),
                            post_path(post),
                            data:  {ref: post.id},
-                           class: "hard_object_link").html_safe
+                           class: "hard_object_link")
+    }
+  end
+
+  def opts_for_comment(comment)
+    {
+      comment_link: link_to(comment.message.title,
+                            post_path(comment.post, anchor: comment.guid),
+                            data:  {ref: comment.id},
+                            class: "hard_object_link")
     }
   end
 
