@@ -120,10 +120,10 @@ DiasporaFederation.configure do |config|
     on :fetch_public_entity do |entity_type, guid|
       entity = Diaspora::Federation::Mappings.model_class_for(entity_type).all_public.find_by(guid: guid)
       case entity
-      when Post
-        Diaspora::Federation::Entities.post(entity)
       when Poll
         Diaspora::Federation::Entities.status_message(entity.status_message)
+      else
+        Diaspora::Federation::Entities.build(entity) if entity
       end
     end
 
