@@ -42,7 +42,7 @@ AppConfig ||= Configurate::Settings.create do
                  File.join(config_dir, "diaspora.toml"),
                  namespace: "configuration", required: false
   else
-    warn "WARNING: diaspora.yml is deprecated and will no longer be read in diaspora 2.0."
+    warn "WARNING: diaspora.yml is deprecated and will no longer be read in diaspora 1.0."
     warn "         Please copy over diaspora.toml.example to diaspora.toml and migrate your settings from diaspora.yml."
 
     add_provider Configurate::Provider::YAML,
