@@ -1,19 +1,23 @@
 # frozen_string_literal: true
 
 And /^I click to hide the post/ do
-  find('.hide_post').click
+  find(".hide_post").click
 end
 
 And /^I click to block the user/ do
-  find('.block_user').click
+  find(".block_user").click
 end
 
 And /^I click to report the post/ do
-  find('.post_report').click
+  find(".post_report").click
+end
+
+And /^I click to report the comment/ do
+  find(".comment_report").click
 end
 
 And /^I click to delete the post/ do
-  find('.remove_post').click
+  find(".remove_post").click
 end
 
 And /^I click to (?:like|unlike) the post/ do
